Extract builtin resources from Terraform
Manual curating a list of resources that terraform provides is a PITA.
This commit provides a script that will extract available resources, all
be it in a somewhat noddy way.

It does the following:
  - pulls the source from GitHub;
  - checks out the provided commit/branch/tag;
  - finds all of the `provider.go` files (they contain resource
    definitions);
  - greps and seds out the resources from their definition.

I've moved the list of resources out to its own file, so we don't have
to merge this list into the generator file.